bug[TW27376]: Fix Branch Manager loading and refresh performance
